+++ typescript/cli/src/consts.ts
@@ -2,6 +2,10 @@ export const MINIMUM_CORE_DEPLOY_GAS = (1e8).toString();
 export const MINIMUM_WARP_DEPLOY_GAS = (1e7).toString();
 export const MINIMUM_TEST_SEND_GAS = (3e5).toString();
 
+
 export enum CRUD_COMMANDS {
   DEPLOY = 'deploy',
 }
+
+export const MINIMUM_AVS_GAS = (3e6).toString();
+