+++ typescript/utils/src/types.ts
@@ -1,3 +1,4 @@
+import type { SignatureLike } from '@ethersproject/bytes';
 import type { BigNumber, ethers } from 'ethers';
 
 export enum ProtocolType {
@@ -28,17 +29,6 @@ export type WithAddress<T> = T & {
   address: Address;
 };
 
-// copied from node_modules/@ethersproject/bytes/src.ts/index.ts
-export type SignatureLike =
-  | {
-      r: string;
-      s?: string;
-      _vs?: string;
-      recoveryParam?: number;
-      v?: number;
-    }
-  | ethers.utils.BytesLike;
-
 export type MerkleProof = {
   branch: ethers.utils.BytesLike[];
   leaf: ethers.utils.BytesLike;
@@ -46,6 +36,13 @@ export type MerkleProof = {
 };
 
 /********* HYPERLANE CORE *********/
+export type Announcement = {
+  mailbox_domain: Domain;
+  mailbox_address: Address;
+  validator: Address;
+  storage_location: string;
+};
+
 export type Checkpoint = {
   root: string;
   index: number; // safe because 2 ** 32 leaves < Number.MAX_VALUE
@@ -53,14 +50,23 @@ export type Checkpoint = {
   merkle_tree_hook_address: Address;
 };
 
+export type CheckpointWithId = {
+  checkpoint: Checkpoint;
+  message_id: HexString;
+};
+
+export { SignatureLike };
+
 /**
  * Shape of a checkpoint in S3 as published by the agent.
  */
 export type S3CheckpointWithId = {
-  value: {
-    checkpoint: Checkpoint;
-    message_id: HexString;
-  };
+  value: CheckpointWithId;
+  signature: SignatureLike;
+};
+
+export type S3Announcement = {
+  value: Announcement;
   signature: SignatureLike;
 };
 
@@ -84,8 +90,10 @@ export type ParsedMessage = {
   version: number;
   nonce: number;
   origin: number;
+  originChain?: string;
   sender: string;
   destination: number;
+  destinationChain?: string;
   recipient: string;
   body: string;
 };
@@ -98,10 +106,6 @@ export type ParsedLegacyMultisigIsmMetadata = {
   signatures: ethers.utils.BytesLike[];
   validators: ethers.utils.BytesLike[];
 };
-
-export enum InterchainSecurityModuleType {
-  MULTISIG = 3,
-}
 
 export type Annotated<T> = T & {
   annotation?: string;